Changed some constants from int to double

+++ src/main/java/frc/robot/Constants.java
@@ -26,13 +26,13 @@ public final class Constants {
     public static final int ELEVATOR_ENCODER = 0;
     public static final int LASERCAN_ID = 0;
 
-    public static final int BOTTOM_POSITION = 0;
-    public static final int TOP_POSITION = 0;
+    public static final double BOTTOM_POSITION = 0;
+    public static final double TOP_POSITION = 0;
 
-    public static final int L1_HEIGHT = 0;
-    public static final int L2_HEIGHT = 0;
-    public static final int L3_HEIGHT = 0;
-    public static final int L4_HEIGHT = 0;
-    public static final int PROCESSOR_HEIGHT = 0;
+    public static final double L1_HEIGHT = 0;
+    public static final double L2_HEIGHT = 0;
+    public static final double L3_HEIGHT = 0;
+    public static final double L4_HEIGHT = 0;
+    public static final double PROCESSOR_HEIGHT = 0;
   }
 }
